Site Editor: add preview options component
Move PreviewOptions and useResizeCanvas to block-editor package.
