+++ packages/block-editor/src/components/preview-options/index.js
@@ -1,24 +1,17 @@
 /**
- * External dependencies
- */
-import { get } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import {
 	Button,
 	Dropdown,
-	Icon,
 	MenuGroup,
 	MenuItem,
 	Path,
 	SVG,
 } from '@wordpress/components';
 import { useSelect, useDispatch } from '@wordpress/data';
-import { PostPreviewButton } from '@wordpress/editor';
 import { __ } from '@wordpress/i18n';
-import { external, check } from '@wordpress/icons';
+import { check } from '@wordpress/icons';
 
 const downArrow = (
 	<SVG
@@ -31,27 +24,15 @@ const downArrow = (
 	</SVG>
 );
 
-export default function PreviewOptions( {
-	forceIsAutosaveable,
-	forcePreviewLink,
-} ) {
+export default function PreviewOptions( { isSaveable = true } ) {
 	const {
 		__experimentalSetPreviewDeviceType: setPreviewDeviceType,
-	} = useDispatch( 'core/edit-post' );
+	} = useDispatch( 'core/block-editor' );
 
 	const deviceType = useSelect( ( select ) => {
-		return select( 'core/edit-post' ).__experimentalGetPreviewDeviceType();
-	}, [] );
-
-	const isSaveable = useSelect( ( select ) => {
-		return select( 'core/editor' ).isEditedPostSaveable();
-	}, [] );
-
-	const isViewable = useSelect( ( select ) => {
-		const { getEditedPostAttribute } = select( 'core/editor' );
-		const { getPostType } = select( 'core' );
-		const postType = getPostType( getEditedPostAttribute( 'type' ) );
-		return get( postType, [ 'viewable' ], false );
+		return select(
+			'core/block-editor'
+		).__experimentalGetPreviewDeviceType();
 	}, [] );
 
 	return (
@@ -95,25 +76,6 @@ export default function PreviewOptions( {
 							{ __( 'Mobile' ) }
 						</MenuItem>
 					</MenuGroup>
-					{ isViewable && (
-						<MenuGroup>
-							<div className="editor-post-preview__grouping-external">
-								<PostPreviewButton
-									className={
-										'editor-post-preview__button-external'
-									}
-									forceIsAutosaveable={ forceIsAutosaveable }
-									forcePreviewLink={ forcePreviewLink }
-									textContent={
-										<>
-											<Icon icon={ external } />
-											{ __( 'Preview in new tab' ) }
-										</>
-									}
-								/>
-							</div>
-						</MenuGroup>
-					) }
 				</>
 			) }
 		/>
