+++ packages/block-editor/src/components/preview-options/style.scss
@@ -59,7 +59,7 @@
 
 @include break-small() {
 	.editor-post-preview {
-		display: none;
+		// display: none;
 	}
 
 	.editor-post-preview__dropdown {
