+++ packages/block-editor/src/store/actions.js
@@ -1004,3 +1004,19 @@ export function toggleBlockHighlight( clientId, isHighlighted ) {
 		isHighlighted,
 	};
 }
+
+/**
+ * Returns an action object used to toggle the width of the editing canvas.
+ * It's marked as experimental because, potentially, we'll need this
+ * in several pages including edit-site.
+ *
+ * @param {string} deviceType
+ *
+ * @return {Object} Action object.
+ */
+export function __experimentalSetPreviewDeviceType( deviceType ) {
+	return {
+		type: 'SET_PREVIEW_DEVICE_TYPE',
+		deviceType,
+	};
+}
