+++ packages/block-editor/src/store/selectors.js
@@ -1633,3 +1633,16 @@ export function didAutomaticChange( state ) {
 export function isBlockHighlighted( state, clientId ) {
 	return state.highlightedBlock === clientId;
 }
+
+/**
+ * Returns the current editing canvas device type.
+ * It's marked as experimental because, potentially, we'll need this
+ * in several pages including edit-site.
+ *
+ * @param {Object} state Global application state.
+ *
+ * @return {string} Device type.
+ */
+export function __experimentalGetPreviewDeviceType( state ) {
+	return state.deviceType;
+}
