+++ packages/edit-post/src/components/header/header-toolbar/index.js
@@ -38,7 +38,7 @@ function HeaderToolbar( { onToggleInserter, isInserterOpen } ) {
 			isTextModeEnabled:
 				select( 'core/edit-post' ).getEditorMode() === 'text',
 			previewDeviceType: select(
-				'core/edit-post'
+				'core/block-editor'
 			).__experimentalGetPreviewDeviceType(),
 		} ),
 		[]
