+++ packages/edit-post/src/editor.native.js
@@ -143,12 +143,12 @@ class Editor extends Component {
 
 export default compose( [
 	withSelect( ( select ) => {
-		const {
-			isFeatureActive,
-			getEditorMode,
-			getPreference,
-			__experimentalGetPreviewDeviceType,
-		} = select( 'core/edit-post' );
+		const { isFeatureActive, getEditorMode, getPreference } = select(
+			'core/edit-post'
+		);
+		const { __experimentalGetPreviewDeviceType } = select(
+			'core/block-editor'
+		);
 		const { getBlockTypes } = select( 'core/blocks' );
 
 		return {
