+++ packages/edit-post/src/store/actions.js
@@ -270,19 +270,3 @@ export function metaBoxUpdatesSuccess() {
 		type: 'META_BOX_UPDATES_SUCCESS',
 	};
 }
-
-/**
- * Returns an action object used to toggle the width of the editing canvas.
- * It's marked as experimental because, potentially, we'll need this
- * in several pages including edit-site.
- *
- * @param {string} deviceType
- *
- * @return {Object} Action object.
- */
-export function __experimentalSetPreviewDeviceType( deviceType ) {
-	return {
-		type: 'SET_PREVIEW_DEVICE_TYPE',
-		deviceType,
-	};
-}
